pyproject.toml
[tool.poetry]
name = "saksa"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "^0.17.3"
wsproto = "^1.0.0"
python-socketio = "^5.3.0"
orjson = "^3.5.3"
aioredis = {extras = ["hiredis"], version = "^2.0.0a1"}
confluent-kafka = "^1.7.0"
python-dotenv = "^0.18.0"
itsdangerous = "^2.0.1"
requests = "^2.27.1"
websocket-client = "^1.2.3"
starlette = "^0.20.0"
scylla-driver = "^3.25.4"
python-multipart = "^0.0.5"
Jinja2 = "^3.1.2"
anyio = "^3.6.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
flake8 = "^3.9.2"
httpx = "^0.23.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | tests/.*/setup.py
)/
'''
